Revert treeifier functionality. Fixes #13
The treeifier was added in #3 as a workaround to a bug in the underlying 'noid' gem. The workaround spawned a discussion of how best to bucket assets that generated a number of possible solutions, none of which seemed satisfactory. The underlying noid bug is now resolved, and the treeifier functionality was released temporarily in AF::Noid 1.0.0 which was yanked, so no one's using the functionality. Therefore, be it resolved that the treeifier is now put to rest, may it rest in peace. (Thanks, @jechols, @hackmastera, and others for helping me through this.)