Bump dependencies and refactor sync minter
Refactor:
- extract AF::Noid::Minter::Base class
- extend noid::minter
- Will be used by DB minter also.

This gets rid of the throwaway new minter per call to `valid?`.
In fact it gets rid of overriding `valid?` at all.

Template method now returns the template object, not the string (though
`my_minter.template.template` gets the string, and the Template object
itself now stringifies to the string, which is probably more obvious).

Rails and sqlite dependencies as part of activerecord-backed minter
work.