Remove db:migrate from the generator
The generator should only be run once (to make the migrations)
but db:migrate must be run in every environment.  This keeps that
consistent and matches the patterns used by other gems that generate
database tables (see Devise for one such example).  As seeds are now
autocreated, that step is no longer necessary either.